requirements.txt
flask
requests
python-instagram
rasa
psycopg2-binary